make functional tests pass
the functional tests were failing because they were trying to compare
the equality of lists, which were apparently not coming back in the
right order.  I've solved this by simply sorting the returned results by
name.

there was also a python3 vs. python2 problem in `ipfsApi.filestream`
that was tripping up these tests.