Update tests and code to pass on with latest IFPS.
go-ipfs v0.4.3rc2 has changed to the IPFS api some.
This commit updates the responses that our tests
expect to match those from this newer API, as well
as fixing the handling for adding strings (which
appears to have changed a lot). It also updates the
README to specify which version of IPFS is being
tested against so that users can make more
informed decisions when using this package.

License: MIT
Signed-off-by: Christopher Waldon <ckwaldon@us.ibm.com>