Fix TravisCI to run functional tests.
TravisCI now installs the latest version of IPFS
and runs the daemon before starting the testing
suite. This means that the functional tests can
now pass, whereas before they "failed to connect"
to a non-running daemon. Unfortunately, TravisCI
will "go get" IPFS, which gives us no control
over the version. We will get whatever the
latest commit is. As soon as go-ipfs v0.4.3 is
finalized and released, the TravisCI config
should be changed to install that instead.

License: MIT
Signed-off-by: Christopher Waldon <ckwaldon@us.ibm.com>