client: add nocopy, raw-leaves options to add cmd
The nocopy command is for adding files by reference to their
location in the filesystem. The nocopy implies raw-leaves.

Requires bug fix in go-ipfs 77cd41a (see issue #4558), ver >0.4.18
(strictly greater).

The nocopy option implies raw-leaves, and the default for raw-leaves
depends on nocopy. This dynamic default is the behavior of the HTTP API
tested via curl.

Adding files to the filestore via nocopy requires that the path
is absolute, not relative. The multipart stream component appens
the Abspath header when it has the absolute path to the file.

In IPFS daemon used for tests, enable the filestore feature.

The tests could leverage the filestore API, but that part of the API is
not yet implemented, so leaving parts of the test commented out.